assignment1_step13

+++ inlupp1/linked_list.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "common.h"
 #include <stdbool.h>
-
+#include <stddef.h>
 /// @brief Creates a new empty list
 /// @param ioopm_eq_function compares two elem_t
 /// @return an empty linked list
@@ -55,7 +55,7 @@ bool ioopm_linked_list_contains(ioopm_list_t *list, elem_t element);
 /// @brief Lookup the number of elements in the linked list in O(1) time
 /// @param list the linked list
 /// @return the number of elements in the list
-int ioopm_linked_list_size(ioopm_list_t *list);
+size_t ioopm_linked_list_size(ioopm_list_t *list);
 
 /// @brief Test whether a list is empty or not
 /// @param list the linked list
